# bedelia-rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bedelia UTEC",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/areasEstudio/{id}": {
            "get": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AreaEstudioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/areasEstudio": {
            "get": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AreaEstudioDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AreaEstudioDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AreaEstudioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/carreras/{id}": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "operationId": "App\\Http\\Controllers\\CarrerasController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarreraDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/carreras/{id}/cursos": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "description": "Devuelve la lista de cursos que conforman la carrera",
                "operationId": "App\\Http\\Controllers\\CarrerasController::obtenerCursosDeCarrera",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/carreras/{id}/previas": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "description": "Devuelve las relaciones de previas entre los cursos que conforman la carrera",
                "operationId": "App\\Http\\Controllers\\CarrerasController::obtenerPreviasEntreCursosDeCarrera",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PreviaDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/carreras/": {
            "post": {
                "tags": [
                    "Carreras"
                ],
                "operationId": "App\\Http\\Controllers\\CarrerasController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarreraCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarreraDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cursos/{id}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cursos": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CursoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/edicionesCurso/{id}/clasesDictada": {
            "post": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Registra una nueva ClaseDictada y la asistencias de los estudiantes",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::Agregar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClaseDictadaDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClaseDictadaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/periodos": {
            "get": {
                "tags": [
                    "Periodos"
                ],
                "operationId": "App\\Http\\Controllers\\PeriodoLectivoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "Devuelve el ultimo anio lectivo ingresado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnioLectivoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Periodos"
                ],
                "operationId": "App\\Http\\Controllers\\PeriodoLectivoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnioLectivoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "No devuelve nada, ya es bastante que funcione xD"
                    }
                }
            }
        },
        "/postulantes/{id}": {
            "get": {
                "tags": [
                    "Postulante"
                ],
                "description": "devuelve una postulacion especifica, incluyendo la persona asociada y su direccion",
                "operationId": "App\\Http\\Controllers\\PostulantesController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulacion",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostulanteDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Postulante"
                ],
                "description": "Elimina la postulacion especifica",
                "operationId": "App\\Http\\Controllers\\PostulantesController::rechazar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/postulantes": {
            "post": {
                "tags": [
                    "Postulante"
                ],
                "description": "Registra en el sistema los datos de un postulante",
                "operationId": "App\\Http\\Controllers\\PostulantesController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostulanteDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostulanteDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sedes/{id}": {
            "get": {
                "tags": [
                    "Sedes"
                ],
                "operationId": "App\\Http\\Controllers\\SedesController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve datos de la sede",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sedes": {
            "post": {
                "tags": [
                    "Sedes"
                ],
                "operationId": "App\\Http\\Controllers\\SedesController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SedeDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve datos de la sede creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tiposCurso/{id}": {
            "get": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoCursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tiposCurso": {
            "get": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoCursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoCursoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoCursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AreaEstudioDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "area": {
                        "type": "string"
                    },
                    "creditos": {
                        "description": "Depende de la carrera",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "CarreraDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "cant_semestres": {
                        "type": "integer"
                    },
                    "areas_estudio": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AreaEstudioDTO"
                        }
                    },
                    "sedes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SedeDTO"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Token de autenticacion conformato como '<token>'",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}